requirements.txt
# Pinned dependencies that cause breakage
Werkzeug==2.2.3
SQLAlchemy==2.0.0

# Runtime dependencies
Flask==2.2.3
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.3
python-dotenv==0.21.1

# Runtime tools
gunicorn==20.1.0
honcho==1.1.0

# Code quality
pylint==2.17.4
flake8==5.0.4
black==23.3.0

# Testing dependencies
nose==1.3.7
pinocchio==0.4.3
factory-boy==3.2.1
coverage==7.1.0
httpie==3.2.1

# Behavior Driven Development
behave==1.2.6
selenium==4.1.0
compare==0.2b0
requests==2.28.2
